server-e2b/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
docker
PyGithub
requests
python-dotenv
supabase==2.6.0
github3.py
PyJWT==2.8.0
pytest==7.4.3
pytest-mock==3.12.0
pydantic==2.5.3
email-validator==2.1.0
e2b
aiohttp==3.9.1